Allow app to be imported
Flask httpdomain needs to import the Flask app. It turns out the the
way to get the path to the configuration file assumed the app was always
being run/imported from the base of the ltd-keeper repo, which isn't
always true, especially for Sphinx doc builds.

This correctly derives a path from the absolute path of the Python
module itself.

For DM-5100.