Outline HTTP API reference documentation
This commit adds a common structure for documenting the core resources
provided by the LTD Keeper api: /products/, /builds/ and /versions/.

The idea is to

- introduce each concept briefly
- provide a summary list of methods (and point out related methods under
  other routes
- use httpdomain's autoflask to provide full documentation for each
  method

More effort is needed to add examples and metadata to the docstrings;
see
http://pythonhosted.org/sphinxcontrib-httpdomain/#roles
for info on adding structured data to the docstrings of routes.

For DM-5100.