Make Build slug an optional input
If the user can't provide a useful slug, one is automatically created
(an incrementing integer series unique to the product).

In the process of doing this, I found that ValidationErrors raised
during the import_data of a Build were creating an uncommited DB state
that tripped up subsequent queries. The solution is to now catch an
exception in the route POST /products/id/build/, rollback the DB, and
re-raise the exception to propagate the error message. This pattern
should be adopted for other POSTs and PUTs.

For DM-4950.