Create and delete CNAMEs with AWS/Route 53
Now depend on boto3 to interact with AWS.

Both the create_cname and delete_cname functions are fairly simplistic
and don't have affordances for geographically-distributed servers and
failovers. We'll see how far we can go with Fastly before we need
extra-fancy CNAME setups.

Also adds logging (with a handler that will be created by the runner
script)

For DM-4951.