Add build git_refs and github_requester fields
This metadata will allow ltd-keeper to detect when a new build is
available for an edition and be able to message the individual who
triggered the documentation build (neither hooks are implemented here).

We're using GitHub handles since that's what will most likely be known
by the ltd-keeper client. Some other tool/service would be needed to
convert GitHub identies into Slack usernames, for example.

For DM-4950.