Remove /v1/ URL prefix for versioning
The consensus seems to be that putting the version into the URL is
un-RESTful. If versioning becomes important then the way to do it will
be to check for version in the Accept header;
e.g. Accept: application/vnd.spire-io.session+json;version=1.0
see http://blog.luisrei.com/articles/rest.html