Add authorization tests for build routes
The upload_build permission covers both the POST to add a new build and
a PATCH method to register the build as uploaded. A separate permission
covers the DELETE method since the person adding a build shouldn't
necessarily be allowed to delete that build (it may be used by an
edition).

- Includes docs for routes
- Includes authorization tests for these secured routes with all
  types of clients (anonymous, and different types of permissions)