Use uWSGI to deploy the Flask app
uWSGI is used as an application server that binds the Flask web app to
Nginx. uwsgi talks to nginx over TCP socket on the 3031 port. I've found
that I don't need to specify the host name at all. I think :3031 is
equivalent to 0.0.0.0:3031

uWSGI connects to the Flask app via the module and object within the
module. Here run.py is where the Flask app object is created.

This is in prep for running dual container pods to separate the nginx
and uwsgi processes.