Add surrogate_key column to Build for Fastly
The surrogate key is intended to be included in the
x-amz-meta-surrogate-key header of all objects in S3 belonging to a
build. This always us to instantly purge a build from Fastly (for
example, when pointing an Edition to a new build).

Since surrogate keys only need to be unique, I'm just using a uuid4
expressed as a hex string. That way the value is always just 32
characters and still always unique.

A surrogate key is automatically issued by a POST
/products/<product>/builds/.