Create a Keeper maintenance pod
Use this to create bare pods for maintenance activities, like running a
migration or initializing a database.

This pod overrides the default command in the uwsgi container to simply
run sleep for 1 day rather than start the uWSGI+Flask app. This way an
operator can log into the container and use run.py init or run.py
migrate, etc..

Also overrides the securityContext to run as root (uuid=0). I found this
necessary because the persistent storage from Google Compute Engine is
owned by root, so a root user needs to initially setup the database
before chown-ing it to the uwsgi user that we normally run.