Add a keeper-pod for single pod deploys
It turns out we can't use the keeper-controller replication controller
since we can't attach persistent storage to a pod managed by a
replication controller. In the meantime, we can deploy a bare pod looked
up to persistent storage. This yaml pod definition does it.

This pod is very similar to the keeper-mgmt-pod, except that all the
hacks to prevent uwsgi from running, and also running as root are
removed. The label is also configured so that the keeper-service load
balancer will send traffic to the pod. Effectively this is a drop-in
replacement for keeper-controller.