Add Nginx SSL Proxy service
Instead of trying to use Kubernetes Ingress to terminate SSL, which I
can't get to work, I've adapted the deployment to use an Nginx SSL
proxy. This setup is based on

https://github.com/GoogleCloudPlatform/kube-jenkins-imager/blob/master/LICENSE

Basically it takes traffic on ports 80 and 443, terminates SSL, and
sends traffic to the keeper service on port 8080. The nginx proxy also
redirects port 80 to use HTTPS instead.

- ssl-proxy-service exposes ports 80 and 443 to the world gives the
  external IP
- keeper-service sets a cluster IP and port that the SSL proxy can send
  traffic to.
- ssl-proxy runs an Nginx container from Google that is configured to
  terminate SSL