Products created with slug and root domains
Previously a Product resource would be created where the domain a
product was served from was specified independently of the slug. With
fastly regex-based redirects we need to tie the slug (and therefore the
directory in the bucket) to the domain a product is served from.

Thus have the user specify the root domain and the root domain that
Fastly services from and then compute the domain specific to the product
itself.

This suggests that perhaps the Fastly service ID, bucket name, fastly
root domain and doc root domain should all be refactored into a single
row of a separate table (i.e., a 'site settings').