Fix query in product/id/builds/ and */editions/
The GET methods for the collections of builds and editions wasn't
properly filtering by product slug. This commit adds true joins between
the Build and Product (and Edition and Product) tables and then filters
the joined Product against the slug supplied in the URL.

D'oh! This should have been tested earlier.