S3 copy can modify surrogate-key header
Previously the s3.copy_directory function would only ensure that
metadata was replicated from the original object to a new object.
Now we have a use case for Editions where the original surrogate-key
needs to be replaced with one standard for that edition.

This patch ensures that old metadata is propagated, while still updating
surrogate-key.

Note that this requires an extra HEAD request on each object before
copying it.