Editions have their own surrogate-keys (migration)
Previously Editions inherited the surrogate key from the source build.
The problem with this is that it permits some degree of unreliability if
the surrogate key for objects actually being served by Fastly for that
edition gets out of sync with the surrogate key for the last used build.
Having a stable surrogate keys for each edition obviates this entire
vector of bugs and makes it easy to manually purge an edition.

- Adds surrogate_key column to editions.
- Editions populate their own surrogate key on their next rebuild (once
  all production editions have this key the column should be set to
  nullable=false).
- Use new s3.copy_directory option to specify a surrogate_key.

In production all editions may need to be rebuilt and then do a Fastly
purge on everything to get everything in sync.