Enable ltd-mason in travis.yml
This enables continuous documentation deployment to LSST the Docs under
the ltd-keeper product name.

NOTE: don't use nit-picky mode with httpdomain. Lots of false positives.

Specify version of ltd-mason for travis as the 0.2.0rc0 in order to
install ltd-mason-travis correctly. Will revert this to mainline
ltd-mason once that is released.