Edition rebuild changes Cache-Control to no-cache
When an edition is rebuild the following headers are set:

Surrogate-Control: max-age=31536000
Cache-Control: no-cache

This ensures that Fastly caches the content in its POPs for a year,
while also telling browsers to check with Fastly whether a client-side
cache is stale based on ETag hashes.

See
https://developers.google.com/web/fundamentals/performance/optimizing-content-efficiency/http-caching
for more information.

In the future it may be a good idea to create more nuanched client-side
caching rules based on content types. For example, CSS and JavaScript
can probably be cached for at least a day without requiring the browser
to re-validate that cache.