Switch to length-255 Unicode columns
This is needed because MySQL has a length limit on varchar columns with
a UNIQUE constraint. Since the text is unicode, there are 3 bytes were
character, meaning that 255 is the maximum allowable length.

See http://stackoverflow.com/q/1814532 for eg
"Specified key was too long; max key length is 767 bytes"

Includes an alembic migration step to to move to Unicode columns