Add Edition.mode field
The column is used to set the tracking strategy for an edition.
Definitions of these strategies are centralized in EditionMode class.

The Edition.set_mode method safely sets this column and validates the
mode integer.

Includes docstring for POST/PATCH/GET /edition routes.

Includes the Edition.mode alembic migration

Note the column is nullable so we will treat Edition.mode = None as
equivalent to EditionMode.GIT_REFS.