Remove Flask SERVER_NAME from k8s deploy
We're no longer using Flask's SERVER_NAME configuration in production to
allow Keeper to respond to requests from proxy URLs, not just the
external URL. This is necessary for the health checks to succeed that
the k8s Ingress resource depends on.

Thus the LTD_KEEPER_URL environment variable is removed from containers
and the `server-name` setting is removed from the ConfigMap.