Configure logstruct loggers
For testing and development we're using a key-value formatter for
usability. In production we'll use a JSON formatter for ease of use with
logstash or stackdriver.

There is duplication of configuration code in the three contexts, but
I'm not sure how to DRY the code without unnecessarily reducing
flexibility later.