Replace Edition.rebuild with set_pending_rebuild
By doing this, the Edition model is no longer responsible for the S3
operations to rebuild an edition.

Instead the workflow is:

1. Edition.set_pending_rebuild() is called by a method like
Edition.import_data() if the build has changed. The new build is set on
the Edition, and the Edition.pending_rebuild field is set.

2. Edition.set_pending_rebuild() adds a celery task to the request's
chain. This async task gets launched by the route handler as the route
closes out.

3. The async task calls Edition.set_rebuild_complete() when the task
finishes to indicate that the DB state is consistent with the S3 state.