Use and configure ProxyFix middleware
When operating behind a trusted proxy, we want to use the X-Forwarded-
headers from the proxy to set the hostname and scheme for the API's URLs.
PROXY_FIX environment lets you toggle the Werkzeug ProxyFix middleware
to do just that (default is "0", or "off").

The TRUST_X_FOR, TRUST_X_PROTO, TRUST_X_HOST, TRUST_X_PORT, and
TRUST_X_PREFIX environment variable settings specify the number of
proxies to trust for the given the header when PROXY_FIX is enabled.