Ensure sort ordering or editions in tests
Some tests were relying on a consistent order of editions from the
/editions/ route. This is mostly because we don't have the ability to
query for editions that match certain specifications, like tracked git
refs.

However, we've found out that sqlite adn postgres and returning
different orderings from query results. This is something that should
probably be fixed through a change to the query or through the schema.

However, this modification ensures a consistent ordering based on the
incrementing IDs of editions, to restore the original intent of the
tests.