Add psycopg2-binary
This enables postgres-based database connections (MySQL is already
supported via PyMySQL).

Note that psycopg2-binary is a wheel package. In theory this could be
problematic if libraries used to compile these wheels conflict with the
runtime system libraries (particularly for ssl). If this becomes an
issue we'll switch to the source-based psycopg2 package and ensure that
the necessary postgres libraries are available in development and CI
environments.

Temporarily deleted flower due to a conflict.